Rename `dynamic-backend` to `probe-backend`
