--- logging-fake-backend/src/main/kotlin/io/spine/logging/dynamic/backend/DynamicBackendFactoryService.kt
@@ -1,65 +1,0 @@
-/*
- * Copyright 2023, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.logging.dynamic.backend
-
-import com.google.auto.service.AutoService
-import io.spine.logging.flogger.backend.LoggerBackend
-import io.spine.logging.flogger.backend.BackendFactory
-
-/**
- * Adapts [DynamicBackendFactory] to be used with Java's
- * [ServiceLoader][java.util.ServiceLoader].
- *
- * Each service should have a public no-args constructor to be used
- * by the loader. Which is not a case for object declarations, as they
- * don't have any constructors at all.
- *
- * Until [KT-25892](https://youtrack.jetbrains.com/issue/KT-25892/Allow-objects-to-be-loaded-via-ServiceLoader-or-similar-API)
- * is implemented, we need an adapter for object services.
- *
- * An explicit delegation can't be applied here because [BackendFactory] is a class.
- * Thus, the following snippet fails to compile:
- *
- * ```
- * @AutoService(BackendFactory::class)
- * class DynamicBackendFactoryService : BackendFactory() by DynamicBackendFactory
- * ```
- */
-@AutoService(BackendFactory::class)
-public class DynamicBackendFactoryService : BackendFactory() {
-
-    /**
-     * Delegates actual backend creation to [DynamicBackendFactory] object.
-     */
-    override fun create(loggingClassName: String): LoggerBackend =
-        DynamicBackendFactory.create(loggingClassName)
-
-    /**
-     * Returns a fully-qualified name of this class.
-     */
-    override fun toString(): String = javaClass.name
-}
